lepsi kontroly u delete

+++ delete.php
@@ -14,6 +14,19 @@ $type = $_REQUEST['type'];
 $id = $_REQUEST['id'];
 
 if ($type=='gifts') {
+
+    //kontrola
+    $giftQuery=$db->prepare('SELECT * FROM gifts WHERE gift_id=:id LIMIT 1;');
+    $giftQuery->execute([':id'=>$_REQUEST['id']]);
+    if ($gift=$giftQuery->fetch(PDO::FETCH_ASSOC)){
+        if ($_SESSION['user_id'] != $gift['gift_for']) {
+            //uzivatel chce smazat cizi darek
+            header('Location: mywishlist.php');
+            exit();
+        }
+    }
+
+    //samotne smazani
     $query = $db->prepare('DELETE FROM gifts WHERE gift_id = :id');
     $query->execute([
         ':id'=>$id
@@ -21,6 +34,10 @@ if ($type=='gifts') {
     header('Location: mywishlist.php');
     exit();
 } elseif ($type=='categories') {
+    if ($_SESSION['role'] != 'administrátor') {
+        header('Location: index.php');
+        exit();
+    }
     $query = $db->prepare('DELETE FROM categories WHERE category_id = :id');
     $query->execute([
         ':id'=>$id
